OSD: fix the usedCapacity (found in testing)

+++ Libraries/AQ_OSD/MAX7456_BattMonitor.h
@@ -65,10 +65,10 @@ void displayVoltage(byte areMotorsArmed) {
     }
 
     if (abs(currentValue)>=100) { // > 10A only display whole amps
-      snprintf(buf,12,"%4dA%5d\24  ", currentValue/10, (int)batteryData[osdBatNo].usedCapacity/1000);
+      snprintf(buf,12,"%4dA%5ld\24  ", currentValue/10, batteryData[osdBatNo].usedCapacity/1000);
     }
     else {
-      snprintf(buf,12,"%c%1d.%1dA%5d\24  ", currentValue<0?'-':' ',abs(currentValue/10),abs(currentValue%10),(int)batteryData[osdBatNo].usedCapacity/1000);
+      snprintf(buf,12,"%c%1d.%1dA%5ld\24  ", currentValue<0?'-':' ',abs(currentValue/10),abs(currentValue%10),batteryData[osdBatNo].usedCapacity/1000);
     }
 
     writeChars( buf, 11, 0, VOLTAGE_ROW+osdBatNo, VOLTAGE_COL+6 );
